doador/requirements.txt
Django>=5.1.15,<7.0
djangorestframework>=3.15.2,<4.0
django-filter>=23.0,<25.0
mysqlclient>=2.2.0,<3.0
python-decouple>=3.8,<4.0
